Refactor Settings Page & Profile Page
Refactor settings page to clean components/hooks and update profile page edit button to actually redirect.
